+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## dbt-databricks 1.11.7 (TBD)
 
+### Features
+
+- Enable Notebook scoped python packages installation
+- 
 ### Fixes
 
+- Fix `workflow_job` Python model submission method failing with dictionary attribute error ([#1360](https://github.com/databricks/dbt-databricks/issues/1360))
 - Fix column order mismatch in microbatch and replace_where incremental strategies by using INSERT BY NAME syntax ([#1338](https://github.com/databricks/dbt-databricks/issues/1338))
 
 ## dbt-databricks 1.11.6 (Mar 10, 2026)